+++ Elements/src/Geometry/Polyline.cs
@@ -807,10 +807,11 @@ namespace Elements.Geometry
         }
 
         /// <summary>
-        /// Split a polyline at the provided points, by inserting new vertices.
+        /// Insert a point into the polyline if it lies along one
+        /// of the polyline's segments.
         /// </summary>
-        /// <param name="points">The points at which to split.</param>
-        /// <param name="closed">Should the polyline be considered closed?</param>
+        /// <param name="points">The points at which to split the polyline.</param>
+        /// <param name="closed">Is the polyline closed?</param>
         protected void Split(IList<Vector3> points, bool closed = false)
         {
             for (var i = 0; i < this.Vertices.Count; i++)