+++ CHANGELOG.md
@@ -18,11 +18,13 @@
 - `Solid.Intersection(SolidOperation a, SolidOperation b)`
 - `SetClassification`
 - `LocalClassification`
+- `ModelArrows`
+- `ModelText`
 
 ### Changed
 
-- Added `Symbols` property to `ContentElement`.
-- Introduce a `SkipCSGUnion` flag on Representation, as a hack to get around CSG failures.
+### Fixed
+- Deduplicate catalog names during code generation.
 
 ## 0.9.2
 
@@ -50,10 +52,19 @@
   - `Polygon.Difference(params Polygon[] polygons)`
   - `Polygon.Union(params Polygon[] polygons)`
 - `Profile.Offset()`
+- Overloads with `maxDistance` parameter for
+  - `Line.ExtendTo(IEnumerable<Line>)`
+  - `Line.ExtendTo(Polyline)`
+  - `Line.ExtendTo(Polygon)`
+  - `Line.ExtendTo(Profile)`
+- Support for DXF from many basic elements.
+
 
 ### Changed
 
 - Some changes to `ContentElement` instance glTF serialization to allow selectability and transformability in the Hypar UI.
+- Added `Symbols` property to `ContentElement`.
+- Introduce a `SkipCSGUnion` flag on Representation, as a hack to get around CSG failures.
 
 ### Fixed